requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
prefect>=2.0.0
docker>=6.0.0
rdflib>=6.0.0
pyyaml>=6.0.0
pytest>=8.0.0
